Unsafe narrowing conversion fixes for Sidebar items
Upstream there has been significant work to convert the use of signed
indexes, to unsigned types, as well as removing the use of -1 as a
sentinel value. This change corrects Brave code enough to get it to work
reasonably well, but there's need for some refacotoring of our UI code
to reflect the same policy.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/f90b2c3f00e48ce1f2e663ce768c6a0684e39c10

commit f90b2c3f00e48ce1f2e663ce768c6a0684e39c10
Author: Peter Kasting <pkasting@chromium.org>
Date:   Mon Jul 25 19:26:58 2022 +0000

    Change View::GetIndexOf() to return absl::optional<size_t>.

    Bug: 1292951

+++ browser/ui/views/sidebar/sidebar_control_view.cc
@@ -183,12 +183,12 @@ std::u16string SidebarControlView::GetTooltipTextFor(
 }
 
 void SidebarControlView::OnItemAdded(const sidebar::SidebarItem& item,
-                                     int index,
+                                     size_t index,
                                      bool user_gesture) {
   UpdateItemAddButtonState();
 }
 
-void SidebarControlView::OnItemRemoved(int index) {
+void SidebarControlView::OnItemRemoved(size_t index) {
   UpdateItemAddButtonState();
 }
